# [hash/shake]: Don't use function pointers