… hash/shake/hashes_generic.go
… hash/shake/xor.go